# apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the CyVerse Discovery Environment: app discovery and launching, analysis management, and iRODS data access.\n\nTo authenticate: open the Authorize dialog and enter your username and password under BasicAuth,\nrun POST /login, then copy the access_token from the response into the BearerAuth value\n(prefixed with \"Bearer \") to call the other endpoints.",
        "title": "formation",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/data/{path}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a JSON listing for collections, or streams raw file contents for data objects\n(offset/limit allow partial reads). With include_metadata=true, AVU metadata is returned\nin X-Datastore-{attribute} response headers, with value and units joined by avu_delimiter.",
                "produces": [
                    "application/json",
                    "application/octet-stream"
                ],
                "tags": [
                    "Data Store"
                ],
                "summary": "Browse a directory or download a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "iRODS path (e.g. cyverse/home/username)",
                        "name": "path",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Starting byte offset when reading a file",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum bytes to read; 0 or absent reads to the end",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Include AVU metadata as response headers",
                        "name": "include_metadata",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": ",",
                        "description": "Separator between value and units in metadata headers",
                        "name": "avu_delimiter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Directory listing (JSON) or raw file contents",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Path not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send a request body to create or update a file; use resource_type=directory with no\nbody to create a collection; send no body to an existing path for a metadata-only\nupdate. Metadata is supplied via X-Datastore-{attribute} request headers (value and\nunits split on avu_delimiter). Swagger UI cannot send arbitrary headers; use curl for\nmetadata operations.",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Store"
                ],
                "summary": "Upload a file, create a directory, or set metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "iRODS path",
                        "name": "path",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "directory"
                        ],
                        "type": "string",
                        "description": "Set to directory to create a collection",
                        "name": "resource_type",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Replace existing AVUs for the attributes being set instead of adding",
                        "name": "replace_metadata",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": ",",
                        "description": "Separator between value and units in metadata headers",
                        "name": "avu_delimiter",
                        "in": "query"
                    },
                    {
                        "description": "Raw file content",
                        "name": "content",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "path, type, and created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ambiguous operation or upload to a directory",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Parent directory not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a data object or collection. Use dry_run=true to preview the outcome without\ndeleting. Deleting a non-empty directory requires recurse=true, in both real and\ndry-run mode.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Store"
                ],
                "summary": "Delete a file or directory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "iRODS path",
                        "name": "path",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Allow deleting non-empty directories",
                        "name": "recurse",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Preview the deletion without executing it",
                        "name": "dry_run",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "path, type, would_delete, deleted, dry_run, and item_count for recursive non-empty directories",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Directory not empty",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Path not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "A Keycloak JWT prefixed with \"Bearer \". Obtain one via POST /login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}